modif

+++ controllers/dashboardDeleteQuestionAnswerCtrl.php
@@ -2,11 +2,16 @@
 require_once(__DIR__ . '/../models/Flash.php');
 require_once(__DIR__ . '/../models/Question.php');
 require_once(__DIR__ . '/../helpers/dd.php');
+require_once(__DIR__ . '/../config/init.php');
 
 
 // Récupération de l'ID
 $id = intval(filter_input(INPUT_GET,'id', FILTER_SANITIZE_NUMBER_INT));
-
+if($_SESSION['user']->role != 2){
+    Flash::setMessage('<i class="me-3 fa-solid fa-ban fa-beat" style="color: #f50031;"></i>  Vous n\'avez pas accès à cette partie du site !', 'danger');
+    header('location: /accueil');
+    die;
+}
 try {
     $result = Question::delete($id);    
     // Test pour vérifier si la suppression à était effectué
